competitions/requirements.txt
openai==1.97.1
pydantic==2.10.6
google-genai==1.56.0
anthropic==0.54.0
pandas==2.1.4
numpy==1.26.4
deepeval==3.4.3
thefuzz==0.22.1
scipy==1.15.2
pydash==4.9.0
json5==0.12.1
regex==2022.3.2
tiktoken==0.12.0
socksio==1.0.0
XlsxWriter==3.2.2

dateparser==1.1.1
tabulate==0.9.0
python-dotenv==1.0.1
